externalise statistics functions
