+++ bundles/org.openhab.core.persistence/src/main/java/org/openhab/core/persistence/extensions/PersistenceExtensions.java
@@ -43,6 +43,7 @@ import org.openhab.core.persistence.QueryablePersistenceService;
 import org.openhab.core.types.State;
 import org.openhab.core.types.TimeSeries;
 import org.openhab.core.types.TypeParser;
+import org.openhab.core.util.Statistics;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -62,7 +63,8 @@ import org.slf4j.LoggerFactory;
  * @author Mark Herwege - Changed return types to State for some interval methods to also return unit
  * @author Mark Herwege - Extended for future dates
  * @author Mark Herwege - lastChange and nextChange methods
- * @author mark Herwege - handle persisted GroupItem with QuantityType
+ * @author Mark Herwege - handle persisted GroupItem with QuantityType
+ * @author Mark Herwege - add median methods
  */
 @Component(immediate = true)
 @NonNullByDefault
@@ -1697,7 +1699,7 @@ public class PersistenceExtensions {
         }
         Unit<?> unit = baseItem instanceof NumberItem numberItem ? numberItem.getUnit() : null;
 
-        ArrayList<BigDecimal> resultList = new ArrayList<>();
+        List<BigDecimal> resultList = new ArrayList<>();
         result.forEach(hi -> {
             DecimalType dtState = getPersistedValue(hi, unit);
             if (dtState != null) {
@@ -1705,132 +1707,15 @@ public class PersistenceExtensions {
             }
         });
 
-        int size = resultList.size();
-        if (size >= 0) {
-            int k = size / 2;
-            BigDecimal median = null;
-            if (size % 2 == 1) {
-                median = QuickSelect.quickSelect(resultList, 0, size - 1, k);
+        BigDecimal median = Statistics.median(resultList);
+        if (median != null) {
+            if (unit != null) {
+                return new QuantityType<>(median, unit);
             } else {
-                median = QuickSelect.quickSelectForcePrevious(resultList, 0, size - 1, k);
-                if (median != null) {
-                    // quickSelect has forced the k-1 element to be in the right place
-                    median = median.add(resultList.get(k - 1)).divide(new BigDecimal(2));
-                }
-            }
-
-            if (median != null) {
-                if (unit != null) {
-                    return new QuantityType<>(median, unit);
-                } else {
-                    return new DecimalType(median);
-                }
+                return new DecimalType(median);
             }
         }
         return null;
-    }
-
-    /**
-     * Class implementing the quickSelect algorithm.
-     * See https://en.wikipedia.org/wiki/Quickselect and https://gist.github.com/unnikked/14c19ba13f6a4bfd00a3
-     */
-    public static class QuickSelect {
-        /**
-         * Find the k-smallest element between indexes l and r in a list.
-         *
-         * @param bdList, list elements will be reordered in place
-         * @param l index of left most element in list to consider
-         * @param r index of right most element in list to consider
-         * @param k
-         * @return
-         */
-        public static @Nullable BigDecimal quickSelect(ArrayList<BigDecimal> bdList, int l, int r, int k) {
-            return quickSelectInternal(bdList, l, r, k, false);
-        }
-
-        /**
-         * Find the k-smallest element between indexes l and r in a list. At the same time, put the element before the
-         * k-smallest element at position k-1 in bdList. This is useful to calculate the median on a list with an uneven
-         * number of elements. The median will then be the sum of the returned value and the element at position k-1
-         * divided by 2.
-         *
-         * @param bdList, list elements will be reordered in place
-         * @param l index of left most element in list to consider
-         * @param r index of right most element in list to consider
-         * @param k
-         * @return
-         */
-        public static @Nullable BigDecimal quickSelectForcePrevious(ArrayList<BigDecimal> bdList, int l, int r, int k) {
-            return quickSelectInternal(bdList, l, r, k, true);
-        }
-
-        /**
-         * Find the k-smallest element between indexes l and r in a list.
-         *
-         * @param bdList, list elements will be reordered in place
-         * @param l index of left most element in list to consider
-         * @param r index of right most element in list to consider
-         * @param k
-         * @param forcePreviousOrder positions the k-1 element in the right place if true
-         * @return
-         */
-        private static @Nullable BigDecimal quickSelectInternal(ArrayList<BigDecimal> bdList, int l, int r, int k,
-                boolean forcePreviousOrder) {
-            if (r < 0) {
-                return null;
-            } else if (r == 0) {
-                return bdList.get(r);
-            }
-
-            int left = l;
-            int right = r;
-            for (;;) {
-                int pivotIndex = randomPivot(left, right);
-                pivotIndex = partition(bdList, left, right, pivotIndex, forcePreviousOrder);
-
-                if (k == pivotIndex) {
-                    return bdList.get(k);
-                } else if (k < pivotIndex) {
-                    right = pivotIndex - 1;
-                } else {
-                    left = pivotIndex + 1;
-                }
-            }
-        }
-
-        private static int partition(ArrayList<BigDecimal> bdList, int left, int right, int pivotIndex,
-                boolean forcePreviousOrder) {
-            BigDecimal pivotValue = bdList.get(pivotIndex);
-            swap(bdList, pivotIndex, right); // Move pivot to end
-            int beforePivotIndex = left;
-            int storeIndex = left;
-            for (int i = left; i < right; i++) {
-                if (bdList.get(i).compareTo(pivotValue) < 0) {
-                    if (forcePreviousOrder && (bdList.get(i).compareTo(bdList.get(beforePivotIndex)) > 0)) {
-                        beforePivotIndex = storeIndex;
-                    }
-                    swap(bdList, storeIndex, i);
-                    storeIndex++;
-                }
-            }
-            swap(bdList, right, storeIndex); // Move pivot to its final place
-            if (forcePreviousOrder && (storeIndex > beforePivotIndex)) {
-                swap(bdList, beforePivotIndex, storeIndex - 1);
-            }
-            return storeIndex;
-        }
-
-        private static void swap(ArrayList<BigDecimal> bdList, int i, int j) {
-            if (i != j) {
-                BigDecimal tmp = bdList.get(i);
-                bdList.set(i, bdList.get(j));
-                bdList.set(j, tmp);
-            }
-        }
-
-        private static int randomPivot(int left, int right) {
-            return left + (int) Math.floor(Math.random() * (right - left + 1));
-        }
     }
 
     /**
